Refactor account mapping tool view model
